# docs/swagger.yaml
definitions:
  controller.AuthResponse:
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
  controller.ErrorResponse:
    properties:
      message:
        example: Service unavailable
        type: string
      status:
        example: error
        type: string
    type: object
  controller.HealthResponse:
    properties:
      message:
        example: Service is healthy
        type: string
      status:
        example: OK
        type: string
    type: object
  controller.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: securepassword123
        type: string
    type: object
  controller.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: securepassword123
        type: string
    type: object
  model.Device:
    properties:
      brand:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      state:
        $ref: '#/definitions/model.DeviceState'
      updated_at:
        type: string
    required:
    - brand
    - name
    - state
    type: object
  model.DeviceState:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - StateInactive
    - StateAvailable
    - StateInUse
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/devices:
    get:
      description: Retrieve devices with optional filters for brand and state
      parameters:
      - description: Filter by brand
        in: query
        name: brand
        type: string
      - description: Filter by state (inactive, available, in-use)
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Device'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Get devices
      tags:
      - devices
    post:
      consumes:
      - application/json
      description: Create a new device with the provided details
      parameters:
      - description: Device details
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/model.Device'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Create a new device
      tags:
      - devices
    put:
      consumes:
      - application/json
      description: Update the details of an existing device
      parameters:
      - description: Updated device details
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/model.Device'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Device'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Update an existing device
      tags:
      - devices
  /api/devices/{id}:
    delete:
      description: Delete a device using its ID
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Delete a device by ID
      tags:
      - devices
    get:
      description: Retrieve the details of a device by its ID
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Device'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Get a device by ID
      tags:
      - devices
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and create session
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /healthz:
    get:
      consumes:
      - application/json
      description: Check if the service is running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.HealthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Health check endpoint
      tags:
      - health
  /readyz:
    get:
      consumes:
      - application/json
      description: Check if the service and all dependencies (database) are ready
        to handle requests
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.HealthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: Readiness check endpoint
      tags:
      - health
swagger: "2.0"